refactor: move basetag to its own dir
