+++ libs/ui-rnative/src/lib/Components/BaseTag/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseTag';
+export * from './types';
