+++ libs/ui-rnative/src/lib/Components/MediaTag/MediaTag.tsx
@@ -1,4 +1,4 @@
-import { BaseTag } from '../Tag/BaseTag';
+import { BaseTag } from '../BaseTag';
 import type { MediaTagProps } from './types';
 
 /**
